Add files via upload
Sample files for a little database I made to track ratings about different attributes of drivers and hostesses of a bus, given a time, date, bus and a specific service. The tool can also produce reports with information that different areas of the company will use for posterior evaluation.

This tool was made to replace a series of paper tables and hand made reports that were used before its existance. There are two different versions, one for long distance services and one for urban services, each having their own attributes and information. Everything is made in spanish language since I'm living in Argentina.